backend/requirements.txt
# Core dependencies
osmnx==1.8.1
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
rtree==1.1.0
networkx==3.2.1

# Data processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Hedonic analysis
statsmodels==0.14.1
scikit-learn==1.3.2

# API framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Data storage
pyarrow==14.0.2
tables==3.9.2

# Caching
diskcache==5.6.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Visualization data export
geojson==3.1.0

# Utilities
tqdm==4.66.1
loguru==0.7.2
click==8.1.7

# HTTP client
httpx==0.26.0
requests==2.31.0
